Updated detector, cleaned up and fixed a minor error
update: had to explicitly set the recursion flag for
XMLUtils.getChildElementsByName to false due to an update in the
library.
clean up: solved all warnings in the file
fix: originally read the loop information from the "Loops" element, not
from the "Loop" element. Tested the change by inspection in the debugger
when reading a file with a loop.