Kontakt File Reading, Renaming During Conversion, Changed Parameters in MPC Export (#18)
* New Detector for Kontakt nki files added, MPCKeygroupCreator improved
(correction of envelope durations & logic to determine the envelope
type).

* Updated detector, cleaned up and fixed a minor error

update: had to explicitly set the recursion flag for
XMLUtils.getChildElementsByName to false due to an update in the
library.
clean up: solved all warnings in the file
fix: originally read the loop information from the "Loops" element, not
from the "Loop" element. Tested the change by inspection in the debugger
when reading a file with a loop.

* Refactored & added NKI NiSS support

Enhanced NKI by support for another former format called NiSS.
Generalized the loading mechanisms so that redundant code between K2
and NiSS is limited to a minimum.
For NiSS there seem to be panning problems at the moment and the loading
of NiSS banks isn't working.

* String properties updated.

* Fixed: panning in NiSS format & support of NKM banks for NiSS

NiSS uses another representation of panning than the K2 format.
Implemented an appropriate conversion.
Found some NKM files containing NiSS programs. Added support for them.

* Support for K2 banks added.

Got access to an old K2 bank NKM file and enhanced the implementation to
load its programs.

---------

Co-authored-by: Philip <Philip@Magrathea.fritz.box>